Fixing
Hey guys I have fixed most of the stuff. Which will make vange work. If it doesnt contact me

+++ SDK/Others.h
@@ -113,7 +113,7 @@ enum ObserverMode_t: int
         OBS_MODE_ROAMING = 6
     };
     
-    enum ItemDefinitionIndex: int
+    enum ItemDefinitionIndex: short
     {
         WEAPON_DEAGLE = 1,
         WEAPON_ELITE = 2,
